--- blaze/client/__init__.py
@@ -1,1 +1,0 @@
-import rarray, requests
