+++ blaze/io/scidb/constructors.py
