+++ blaze/io/scidb/error.py
